+++ controller/socmed.js
@@ -1,16 +1,15 @@
-const {Comment, SocialMedia, User} = require("../models");
-class SocialMediaC{
+const {SocialMedia, User} = require("../models");
 
-    // Akses Sudah Login
-    static async getAllSocialMedias (req, res) {
+class SocialMediaC {
+
+    static async getAllSocialMedias(req, res) {
         try {
             const data = await SocialMedia.findAll({
-                include: User
+                include: [{model: User, attributes: ["id", "username", "profile_image_url"]}]
             })
 
             res.status(200).json(data)
         } catch (error) {
-            console.log(error)
             res.status(500).json(error)
         }
     }
@@ -18,16 +17,16 @@ class SocialMediaC{
     // Akses susuai ID
     static async getSocialMediaById(req, res) {
         try {
-            const { SocialMediaId } = req.params
-            const userData = req.userData
+            const {SocialMediaId} = req.params
+            const user = req.user
             const data = await SocialMedia.findOne({
                 where: {
-                    id : parseInt(SocialMediaId),
-                    UserId: userData.id
+                    id: parseInt(SocialMediaId),
+                    UserId: user.id
                 }
             })
 
-            if(!data) {
+            if (!data) {
                 throw {
                     code: 404,
                     message: "SocialMedia not found!"
@@ -36,52 +35,65 @@ class SocialMediaC{
 
             res.status(200).json(data)
         } catch (error) {
-            console.log(error)
             res.status(error.code || 500).json(error.message)
         }
     }
 
     // Add SocialMedias
-    static async addSocialMedia (req, res) {
+    static async addSocialMedia(req, res) {
         try {
             const {
                 name,
                 social_media_url
             } = req.body
 
-            const userData = req.userData
-            // console.log(userData, "<<userdata")
+            const user = req.user
 
             const data = await SocialMedia.create({
                 name,
                 social_media_url,
-                UserId: userData.id
+                UserId: user.id
+            }).catch(() => {
+                throw {
+                    code: 400
+                }
             });
-           
+
             res.status(201).json(data)
         } catch (error) {
-            res.status(500).json(error)
+            res.sendStatus(error.code || 500)
         }
     }
 
     // Update SocialMedia
-    static async updateSocialMedia (req, res) {
+    static async updateSocialMedia(req, res) {
         try {
             const {
-                name,
-                social_media_url
+                name = null,
+                social_media_url = null
             } = req.body
 
-            const { SocialMediaId } = req.params
+            const {SocialMediaId} = req.params
+            const foundSocmed = await SocialMedia.findByPk(parseInt(SocialMediaId),{});
+            if(!foundSocmed){
+                throw {
+                    code: 404
+                }
+            }
+            if(req.user.id !== parseInt(foundSocmed.UserId)){
+                throw {
+                    code: 403
+                }
+            }
 
             const data = await SocialMedia.update({
                 name,
                 social_media_url
             }, {
                 where: {
-                    id : parseInt(SocialMediaId)
+                    id: parseInt(SocialMediaId)
                 },
-                returning : true
+                returning: true
             })
 
             if (!data[0]) {
@@ -93,7 +105,6 @@ class SocialMediaC{
 
             res.status(201).json(data)
         } catch (error) {
-            console.log(error)
             res.status(error.code || 500).json(error.message)
         }
     }
@@ -101,10 +112,21 @@ class SocialMediaC{
     //Delete SocialMedia By ID
     static async deleteSocialMediaById(req, res) {
         try {
-            const { SocialMediaId } = req.params
+            const {SocialMediaId} = req.params
+            const foundSocmed = await SocialMedia.findByPk(parseInt(SocialMediaId),{});
+            if(!foundSocmed){
+                throw {
+                    code: 404
+                }
+            }
+            if(req.user.id !== parseInt(foundSocmed.UserId)){
+                throw {
+                    code: 403
+                }
+            }
             const data = await SocialMedia.destroy({
-                where:{
-                    id : parseInt(SocialMediaId)
+                where: {
+                    id: parseInt(SocialMediaId)
                 }
             })
             if (!data) {
@@ -116,9 +138,9 @@ class SocialMediaC{
             res.status(200).json("Your SocialMedia has been successfully deleted")
 
         } catch (error) {
-            console.log(error)
-            res.status(error.code || 500).json(error.message)
+            res.sendStatus(error.code || 500)
         }
     }
 }
+
 module.exports = SocialMediaC;
